rest_api/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.11
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
